requirements.txt
pygame==2.1.2
python-benedict==0.27.0
pytest==7.2.0